+++ .dotnet/examples/Chat/Example10_ChatWithAudioAsync.cs
@@ -27,23 +27,23 @@ public partial class ChatExamples
         // Output audio is requested by configuring AudioOptions on ChatCompletionOptions
         ChatCompletionOptions options = new()
         {
-            AudioOptions = new(ChatResponseVoice.Alloy, ChatOutputAudioFormat.Mp3),
+            AudioOptions = new(ChatOutputAudioVoice.Alloy, ChatOutputAudioFormat.Mp3),
         };
 
         ChatCompletion completion = await client.CompleteChatAsync(messages, options);
 
         async Task PrintAudioContentAsync()
         {
-            if (completion.ResponseAudio is ChatResponseAudio responseAudio)
+            if (completion.OutputAudio is ChatOutputAudio outputAudio)
             {
-                Console.WriteLine($"Response audio transcript: {responseAudio.Transcript}");
-                string outputFilePath = $"{responseAudio.Id}.mp3";
+                Console.WriteLine($"Response audio transcript: {outputAudio.Transcript}");
+                string outputFilePath = $"{outputAudio.Id}.mp3";
                 using (FileStream outputFileStream = File.OpenWrite(outputFilePath))
                 {
-                    await outputFileStream.WriteAsync(responseAudio.Data);
+                    await outputFileStream.WriteAsync(outputAudio.Data);
                 }
                 Console.WriteLine($"Response audio written to file: {outputFilePath}");
-                Console.WriteLine($"Valid on followup requests until: {responseAudio.ExpiresAt}");
+                Console.WriteLine($"Valid on followup requests until: {outputAudio.ExpiresAt}");
             }
         }